# src/tsoa/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"UserListResponse": {
				"properties": {
					"users": {
						"items": {
							"properties": {
								"profileUrl": {
									"type": "string"
								},
								"profileIdx": {
									"type": "string"
								},
								"username": {
									"type": "string"
								},
								"id": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"profileUrl",
								"profileIdx",
								"username",
								"id"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"users"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateImageReq": {
				"properties": {
					"imageUrl": {
						"type": "string"
					}
				},
				"required": [
					"imageUrl"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateIntroReq": {
				"properties": {
					"intro": {
						"type": "string"
					},
					"newIntro": {
						"type": "string"
					}
				},
				"required": [
					"intro",
					"newIntro"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateOrCreateFandomReq": {
				"properties": {
					"newFandomIds": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					}
				},
				"required": [
					"newFandomIds"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ScheduleResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"user_id": {
						"type": "number",
						"format": "double"
					},
					"content": {
						"type": "string"
					},
					"memo": {
						"type": "string",
						"nullable": true
					},
					"isDone": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"user_id",
					"content",
					"memo",
					"isDone"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LikePostResponse": {
				"properties": {
					"liked": {
						"type": "boolean"
					}
				},
				"required": [
					"liked"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LikedUserPreviewDto": {
				"properties": {
					"userId": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"profileImageUrl": {
						"type": "string",
						"nullable": true
					},
					"isFollowing": {
						"type": "boolean"
					}
				},
				"required": [
					"userId",
					"username",
					"profileImageUrl",
					"isFollowing"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LikedUsersCursorResponse": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/LikedUserPreviewDto"
						},
						"type": "array"
					},
					"nextCursor": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"hasNext": {
						"type": "boolean"
					}
				},
				"required": [
					"items",
					"nextCursor",
					"hasNext"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PostResponseDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"likeCount": {
						"type": "number",
						"format": "double"
					},
					"commentCount": {
						"type": "number",
						"format": "double"
					},
					"imageCount": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"title",
					"content",
					"likeCount",
					"commentCount"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreatePostDTO": {
				"properties": {
					"title": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"image_url": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"title",
					"content",
					"image_url"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "ovly-express",
		"version": "1.0.0",
		"description": "express mjs ts",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "ripetangerine",
			"url": "kwak jk"
		}
	},
	"paths": {
		"/users/{viewerId}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "viewerId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/search": {
			"get": {
				"operationId": "SearchUsersByName",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserListResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "keyword",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/me/image": {
			"patch": {
				"operationId": "UpdateProfileImage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateImageReq"
							}
						}
					}
				}
			}
		},
		"/users/me/intro": {
			"patch": {
				"operationId": "UpdateProfileIntro",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateIntroReq"
							}
						}
					}
				}
			}
		},
		"/users/me/fandom": {
			"post": {
				"operationId": "UpdateFandom",
				"responses": {
					"201": {
						"description": "Create"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateOrCreateFandomReq"
							}
						}
					}
				}
			}
		},
...
		"/schedule": {
			"post": {
				"operationId": "CreateSchedule",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ScheduleResponse"
								}
							}
						}
					}
				},
				"tags": [
					"schedule"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"get": {
				"operationId": "GetScheduleByDate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ScheduleResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"schedule"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "targetDate",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/schedule/{scheduleId}": {
			"put": {
				"operationId": "UpdateSchedule",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ScheduleResponse"
								}
							}
						}
					}
				},
				"tags": [
					"schedule"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "scheduleId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/posts/{postId}/likes": {
			"post": {
				"operationId": "ToggleLikePost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LikePostResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Post"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "postId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"get": {
				"operationId": "GetLikedUsersAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LikedUsersCursorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Post"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "postId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"default": null,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/posts/{postId}/hide": {
			"post": {
				"operationId": "HidePost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Post"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "postId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/posts/{postId}/report": {
			"post": {
				"operationId": "ReportPost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Post"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "postId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"reason": {
										"type": "string"
									}
								},
								"required": [
									"reason"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/posts": {
			"post": {
				"operationId": "CreatePost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/PostResponseDTO"
										},
										{
											"properties": {
												"message": {}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Post"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreatePostDTO"
							}
						}
					}
				}
			}
		},
		"/posts/ovly/{type}": {
			"get": {
				"operationId": "GetFollowingPostAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Post"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "type",
						"required": false,
						"schema": {
							"default": "suggest",
							"type": "string",
							"enum": [
								"following",
								"suggest"
							]
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"default": null,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/posts/{postId}": {
			"get": {
				"operationId": "GetPostDetail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Post"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "postId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdatePost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Post"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "postId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"content": {
										"type": "string"
									},
									"title": {
										"type": "string"
									}
								},
								"required": [
									"content",
									"title"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeletePost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Post"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "postId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}